ルーティング制御をrouter.js、stateの変更をmutationsに記述
